requirements.txt
fastapi
uvicorn
speedtest-cli
psutil
pyserial
